+++ guillotina/fields/__init__.py
@@ -1,2 +1,3 @@
 from guillotina.fields.annotation import BucketListField  # noqa
+from guillotina.fields.dynamic import DynamicField  # noqa
 from guillotina.fields.patch import PatchField  # noqa